@@ -41,7 +41,7 @@ class PartRecord extends Model
      *
      * @var string[]
      */
-    public $sortIncludeColumns = [
+    public array $sortIncludeColumns = [
         'category.name',
         'vendor.name',
         'channel.name',
@@ -54,7 +54,7 @@ class PartRecord extends Model
      *
      * @var string[]
      */
-    public $sortExceptColumns = [
+    public array $sortExceptColumns = [
         'category_id',
         'vendor_id',
         'purchased_channel_id',
@@ -152,6 +152,7 @@ class PartRecord extends Model
         try {
             return parent::delete();
         } catch (Exception $e) {
-        } // TODO: Change the autogenerated stub
+
+        }
     }
 }
