@@ -5,7 +5,6 @@ namespace App\Models;
 use Dcat\Admin\Traits\HasDateTimeFormatter;
 use Dcat\Admin\Traits\ModelTree;
 use Exception;
-use Illuminate\Database\Eloquent\HigherOrderBuilderProxy;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\HasOne;
 use Illuminate\Database\Eloquent\SoftDeletes;
@@ -66,9 +65,9 @@ class ConsumableRecord extends Model
     /**
      * 耗材总数.
      *
-     * @return HigherOrderBuilderProxy|int|mixed
+     * @return float
      */
-    public function allCounts()
+    public function allCounts(): float
     {
         $consumable_track = $this->track()->first();
         if (empty($consumable_track)) {
@@ -79,7 +78,8 @@ class ConsumableRecord extends Model
     }
 
     /**
-     * 耗材记录有很多耗材履历.
+     * 耗材记录有一个履历.
+     *
      * @return HasOne
      */
     public function track(): HasOne
@@ -97,6 +97,7 @@ class ConsumableRecord extends Model
         try {
             return parent::delete();
         } catch (Exception $e) {
-        } // TODO: Change the autogenerated stub
+
+        }
     }
 }
