@@ -218,7 +218,7 @@ class DeviceRecord extends Model
     /**
      * 判断设备是否借用状态
      *
-     * @return null|bool
+     * @return bool
      */
     public function isLend(): bool
     {
@@ -281,6 +281,7 @@ class DeviceRecord extends Model
         try {
             return parent::delete();
         } catch (Exception $e) {
-        } // TODO: Change the autogenerated stub
+
+        }
     }
 }
