@@ -101,6 +101,7 @@ class ServiceRecord extends Model
         try {
             return parent::delete();
         } catch (Exception $e) {
-        } // TODO: Change the autogenerated stub
+            
+        }
     }
 }
