@@ -108,7 +108,7 @@ class SoftwareRecord extends Model
      *
      * @return int|string
      */
-    public function leftCounts()
+    public function leftCounts(): int|string
     {
         $used = $this->track()->count();
         $counts = $this->counts;
@@ -139,6 +139,7 @@ class SoftwareRecord extends Model
         try {
             return parent::delete();
         } catch (Exception $e) {
-        } // TODO: Change the autogenerated stub
+            
+        }
     }
 }
